Alots of changes
-out dir
-production dir
+test dir
+Portfolio export method
+Equity export method

+++ src/model/Account.java
@@ -7,7 +7,7 @@ public class Account {
 
     private int id;
     public String name;
-    private int amount;
+    private double amount;
 
     public Account(String name, int id, int amount) {
         this.id = id;
@@ -16,10 +16,10 @@ public class Account {
     }
 
     public int getId() { return id; }
-    public int getAmount() { return amount; }
+    public double getAmount() { return amount; }
     public String getName() { return name; }
     // withdraws money from the account. Returns amount left.
-    public double withdraw(int withdrawal) {
+    public double withdraw(double withdrawal) {
 	if(this.amount - withdrawal >= 0)
 	    return this.amount -= withdrawal; // don't remember if this works in java lol. I may or may not
 	// be drunk coding right now...
@@ -27,7 +27,7 @@ public class Account {
 	    return -1; 
     }
     // deposits money into the account.
-    public int deposit(int deposit) {
+    public double deposit(double deposit) {
 	this.amount += deposit;
 	return this.amount;
     }
